pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckduckgo-search>=8.0.1",
    "fastapi>=0.115.12",
    "langchain>=0.3.25",
    "langchain-community>=0.3.23",
    "langchain-core>=0.3.58",
    "langchain-openai>=0.3.16",
    "numpy>=2.2.5",
    "openai>=1.77.0",
    "pydantic>=2.11.4",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "python-dotenv>=1.1.0",
    "sounddevice>=0.5.1",
    "uvicorn>=0.34.2",
    "websockets==11.0.3",
]
